# receiver/openapi.yml
openapi: '3.0.0'
info:
  description: This API receives operational data from Wi-Fi smart-plugs deployed in customer sites.
  version: '1.0.0'
  title: Smart-plug Operational Data API
  contact:
    email: 'tdezwart@my.bcit.ca'

paths:
  /health:
    get:
      summary: Returns 200 OK
      description: A simple endpoint to check if the API is running at all
      operationId: app.health
      responses:
        '200':
          description: API is alive and running

  /plug-data/energy-consumed:
    post:
      summary: Reports a batch of energy consumption readings
      operationId: app.report_energy_consumption_readings
      description: Endpoint to receive energy consumption readings from Wi-Fi smart-plugs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnergyConsumedReadingBatch'
      responses:
        '201':
          description: Energy consumption readings successfully received
        '400':
          description: 'Bad request, e.g. malformed JSON or missing required fields'

  /plug-data/internal-temp:
    post:
      summary: Reports a batch of internal temperature readings
      operationId: app.report_internal_temp_readings
      description: Endpoint to receive internal temperature readings from Wi-Fi smart-plugs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalTempReadingBatch'
      responses:
        '201':
          description: Internal temperature readings successfully received
        '400':
          description: 'Bad request, e.g. malformed JSON or missing required fields'

components:
  schemas:
    EnergyConsumedReadingBatch:
      type: object
      required:
        - plug_id
        - plug_uptime
        - report_timestamp
        - readings
      properties:
        plug_id:
          type: string
          format: uuid
          description: Unique identifier for the smart-plug
          example: '9eb73489-5ada-4359-959e-d6bec07c38f7'
        plug_country:
          type: string
          description: ISO 3166-1 alpha-2 country code where the smart-plug is deployed
          example: 'CA'
        plug_uptime:
          type: integer
          description: Uptime of the smart-plug in seconds
          example: 86400
        report_timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the batch was sent
          example: '2025-01-13T12:34:59Z'
        readings:
          type: array
          items:
            $ref: '#/components/schemas/EnergyConsumedReading'

    EnergyConsumedReading:
      type: object
      required:
        - energy_consumed_watt_minutes
        - switch_state
        - reading_timestamp
      properties:
        energy_consumed_watt_minutes:
          type: integer
          description: Energy consumed over the last minute in watt-minutes
          example: 15
        switch_state:
          type: string
          description: State of the smart-plug switch during the reading.
          enum:
            - 'ON'
            - 'OFF'
          example: 'ON'
        reading_timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the reading was taken
          example: '2025-01-13T12:34:59Z'
    
    InternalTempReadingBatch:
      type: object
      required:
        - plug_id
        - plug_uptime
        - report_timestamp
        - readings
      properties:
        plug_id:
          type: string
          format: uuid
          description: Unique identifier for the smart-plug
          example: '9eb73489-5ada-4359-959e-d6bec07c38f7'
        plug_country:
          type: string
          description: ISO 3166-1 alpha-2 country code where the smart-plug is deployed
          example: 'CA'
        plug_uptime:
          type: integer
          description: Uptime of the smart-plug in seconds
          example: 86400
        report_timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the batch was sent
          example: '2025-01-13T12:34:59Z'
        readings:
          type: array
          items:
            $ref: '#/components/schemas/InternalTempReading'

    InternalTempReading:
      type: object
      required:
        - internal_temp_celsius
        - thermal_status
        - reading_timestamp
      properties:
        internal_temp_celsius:
          type: number
          description: Internal temperature in degrees Celsius
          example: 30.5
        thermal_status:
          type: string
          description: Thermal safety status of the smart-plug during the reading, based on MCU/SoC temperature thresholds
          enum:
            - NOT_AVAILABLE
            - NORMAL
            - WARNING
            - CRITICAL
          example: 'NORMAL'
        reading_timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp of when the reading was taken
          example: '2025-01-13T12:34:59Z'